+++ src/main/java/com/bookshop/sachservice/repository/SachRepository.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
-public interface SachRepository extends MongoRepository<Sach, Integer> {
+public interface SachRepository extends MongoRepository<Sach, Long> {
 
     Optional<Sach> findByTenAndTacGia(String s, String tg);
     @Query("{'$and':  [{'ten': {'$regex': ?0, '$options': 'i'}}, " +
